+++ resources/js/pages/Adminka/pages/Serials/index.js
