+++ resources/js/pages/Example/index.js
