+++ resources/js/pages/Favorites.js
@@ -1,0 +1,8 @@
+//страница Избранное
+import { Button } from "react-bootstrap";
+export const Favorites = () => {
+
+    return (
+
+    );
+};
