+++ resources/js/pages/NotFound/index.js
