+++ resources/js/pages/SignUp/index.js
