--- resources/js/routes/Favorites/index.js
@@ -1,13 +1,0 @@
-//страница Избранное
-import { Button } from "react-bootstrap";
-export const Favorites = () => {
-
-    return (
-        <div>
-            <h2>Избранное</h2>
-            <Button variant="primary">Primary</Button>{' '}
-            <Button variant="secondary">Secondary</Button>{' '}
-            <Button variant="secondary">Secondary</Button>{' '}
-        </div>
-    );
-};
